Extract AtlasPage, MapPage, TagsPage, tagGraph to frontend plugins
